requirements.txt
Flask==2.0.2
google-generativeai==0.1.0
werkzeug==2.0.2